create user dto, adapter to repository in sqlite3, security and service

+++ workflows_documents_user-service/internal/model/user.go
@@ -45,7 +45,7 @@ type (
 		actions []userMod
 	}
 	User struct {
-		ID        int    `db:"id"`
+		ID        int64  `db:"id"`
 		Name      string `db:"name"`
 		LastName  string `db:"last_name"`
 		Email     string `db:"email"`
